MythRAOPConnection: fix Coverity ID 1418589 Unintentional integer overflow
The expression's value may not be what the programmer intended, because the
expression is evaluated using a narrow (i.e. few bits) integer type.

In MythRAOPConnection::ProcessAudio(): An integer overflow occurs, with the
result converted to a wider integer type.

+++ mythtv/libs/libmythtv/AirPlay/mythraopconnection.cpp
@@ -699,7 +699,7 @@ void MythRAOPConnection::ProcessAudio()
 
     // Also make sure m_audioQueue never goes to less than 1/3 of the RDP stream
     // total latency, this should gives us enough time to receive missed packets
-    int64_t queue = framesToMs(m_audioQueue.size() * m_framesPerPacket);
+    int64_t queue = framesToMs((uint64_t)m_audioQueue.size() * m_framesPerPacket);
     if (queue < m_bufferLength / 3)
         return;
